fix size_t and some lint

+++ include/api.h
@@ -4,7 +4,8 @@
 extern "C" {
 #endif
 
-#include <cstdint>
+#include <stdint.h>
+#include <stddef.h>
 
 #define CTI_EXPORT
 #define OUT
@@ -20,11 +21,11 @@ typedef union {
   const char* v_str;
 } packedvalue_handle;
 
-CTI_EXPORT int CTIRegistryListNames(const char* tag, OUT int* ret_size, OUT const char*** ret_names);
+CTI_EXPORT int CTIRegistryListNames(const char* tag, OUT size_t* ret_size, OUT const char*** ret_names);
 
 CTI_EXPORT int CTIRegistryGet(const char* tag, const char* name, OUT func_handle* ret_handle);
 
-CTI_EXPORT int CTIPackedFuncCall(const void* handle, int num_args, const unsigned* type_codes, const packedvalue_handle* values,
+CTI_EXPORT int CTIPackedFuncCall(const void* handle, size_t num_args, const unsigned* type_codes, const packedvalue_handle* values,
     OUT unsigned* ret_type, OUT packedvalue_handle* ret_val);
 
 #ifdef __cplusplus
